apps/clinica/serializers.py
from rest_framework import serializers
from .models import Servicio

class ServicioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Servicio
        fields = ['id','nombreservicio', 'descripcion', 'costo']